Bugfix: Objectmanager delete download folder
Apache log is showning: "rm: cannot remove `download': No such file or
directory" objectmanager constructor was trying to delete this folder
even if it does not exist.

+++ fabui/application/modules/objectmanager/objectmanager.php
@@ -13,7 +13,9 @@ class Objectmanager extends Module {
         }
 
 		$this -> lang -> load($_SESSION['language']['name'], $_SESSION['language']['name']);
-		shell_exec('sudo rm -r download');
+		if (file_exists('download')) {
+			shell_exec('sudo rm -r download');
+		}
 
 	}
